Try to remove the `.trashcan` directory
We disable the trash translators some time ago, but many volumes still
have their `.trashcan` directory. This leads to problems because it will
show up when you mount the volume. MySQL for example thinks it is a
database and tries to open the ".trashcan" DB (and fails)

With this change, the recycler tries to remove `.trashcan` directory,
but does not complain if it fails.