Switch to cron job object
Start using the recently added "oneshot" mode to run the recycler as
CronJob object in OpenShift. Doing so has the advantage that no pod
needs to be running while the recycler would sleep anyway. Users can
tweak the schedule to their liking (i.e. once a day). Error recovery
is also improved as pods are automatically terminated after their
deadline is exceeded.

With the use of "backoffLimit" in the job object the OpenShift version
dependency needs to be upped to 3.9.