Support getting Gluster servers from endpoint objects
When Gluster volumes are hosted on separate Gluster clusters it's no
longer possible to specify all storage servers in a single list, at
least not without extending the format. What this change does instead is
to use the endpoint object named in the PV object to retrieve the
Gluster IP addresses.