Explicitly set pod image for external image
When an external image is used, i.e. from Docker Hub, the image trigger
may not be invoked immediately. In such cases the pod image would remain
set to a single space (" ") which isn't a valid image. Instead this
change reconfigures the pod image to the external image, if used, and
relies on the trigger for later updates.