WIP REPLACE/PUT? Document race condition in recreating volumes
There is a race condition when persistent volume objects are recreated.
A well-timed SIGTERM from the pod being terminated could lead to objects
being deleted but not recreated. Disabling signals is neigh impossible
in shell scripts. The only sane solution to this problem will be to use
a programming language capable of handling/deferring signals.