Add timeouts to content deletion, overall recycling
A job pod is forcefully terminated after it's been running for longer
than its "spec.activeDeadlineSeconds", usually 2h. There are
environments where it's configured to even shorter durations, i.e. 30
minutes.

With this change the OpenShift/Kubernetes API us used to retrieve the
deadline duration. The purpose is to always know the actual deadline
within the pod. In combination with the also added timeout on volume
content removal it's possible to abort a recycle run in time before the
pod is terminated.